requirements.txt
fitz
pytesseract
Pillow
pandas
re
os